Moved to using a sparse tensor for the targets and float32 for the source

+++ data/loader.py
@@ -78,11 +78,11 @@ class WormDataset(Dataset):
             source_image = np.array(hdul).astype("int16")
             source_image = nd.interpolation.zoom(source_image, zoom=0.5)
             source_image = source_image.astype(float) / 4095.0
-            source_image = source_image.astype(np.float32)
+            source_image = source_image.astype(np.float16)
             source_image = np.expand_dims(source_image, axis=0)
             # Divide by the maximum possible in order to normalise the input. Should help with
             # exploding gradients and optimisation.
-            source_image = torch.tensor(source_image, dtype=torch.float32, device=self.device)
+            source_image = torch.tensor(source_image, dtype=torch.float16, device=self.device)
 
         img_path = os.path.join(self.img_dir, self.img_targets.iloc[idx, 1])
 
@@ -92,7 +92,7 @@ class WormDataset(Dataset):
             target_asi = nd.interpolation.zoom(target_asi, zoom=0.5)
             target_asi = binaryise(target_asi)
             target_asi = np.expand_dims(target_asi, axis=0)
-            target_asi = make_sparse(target_asi, torch.int8, self.device)
+            target_asi = make_sparse(target_asi, torch.float16, self.device)
    
         img_path = os.path.join(self.img_dir, self.img_targets.iloc[idx, 2])
         
@@ -102,7 +102,7 @@ class WormDataset(Dataset):
             target_asj = nd.interpolation.zoom(target_asj, zoom=0.5)
             target_asj = binaryise(target_asj)
             target_asj = np.expand_dims(target_asj, axis=0)
-            target_asj = make_sparse(target_asj, torch.int8, self.device)
+            target_asj = make_sparse(target_asj, torch.float16, self.device)
 
         if self.transform:
             source_image = self.transform(source_image)
